+++ src/calmjs/webpack/base.py
@@ -5,10 +5,17 @@ Base classes and constants.
 
 from __future__ import unicode_literals
 
+from collections import namedtuple
+
 # keys
 
 # enable calmjs compatibility - i.e. the dynamic import feature
 CALMJS_COMPAT = 'calmjs_compat'
+# the map from a module name to the loader needed; used by the various
+# functions and methods in the loaderplugin module
+# see definition of WebpackModuleLoaderRegistryKey later
+CALMJS_WEBPACK_MODNAME_LOADER_MAP = 'calmjs_webpack_modname_loader_map'
+
 # The spec key for storing the base webpack configuration.
 WEBPACK_CONFIG = 'webpack_config'
 # The key for the webpack.output.library
@@ -23,6 +30,8 @@ WEBPACK_ENTRY_POINT = 'webpack_entry_point'
 # For webpack loaderplugin integration - this is the spec key - note that
 # this is NOT for webpack plugins which are a separate type of things
 WEBPACK_LOADERPLUGINS = 'webpack_loaderplugins'
+# for the module.rules section; used by loaderplugin module
+WEBPACK_MODULE_RULES = 'webpack_module_rules'
 # for the configuration in webpack config
 WEBPACK_RESOLVELOADER_ALIAS = 'webpack_resolveloader_alias'
 
@@ -67,3 +76,11 @@ DEFAULT_BOOTSTRAP_EXPORT_CONFIG = {
     "root": DEFAULT_BOOTSTRAP_EXPORT,
     "amd": DEFAULT_BOOTSTRAP_EXPORT,
 }
+
+
+# due to webpack specific requirements, a special type for the key is
+# needed for the WebpackModuleLoaderRegistry such that the correct
+# handling mechanism may be done.
+CALMJS_WEBPACK_MODULE_LOADER_SUFFIX = '.webpackloader'
+WebpackModuleLoaderRegistryKey = namedtuple(
+    'WebpackModuleLoaderRegistryKey', ['loader', 'modname'])