+++ src/calmjs/webpack/tests/test_integration.py
@@ -17,6 +17,7 @@ from calmjs.cli import node
 from calmjs import runtime
 from calmjs.utils import pretty_logging
 from calmjs.registry import get as get_registry
+from calmjs.loaderplugin import LoaderPluginRegistry
 
 from calmjs.parse.parsers.es5 import parse
 from calmjs.parse.walkers import ReprWalker
@@ -30,14 +31,18 @@ from calmjs.webpack import toolchain
 from calmjs.webpack import cli
 from calmjs.webpack import exc
 from calmjs.webpack import interrogation
+from calmjs.webpack import loaderplugin
 from calmjs.webpack.base import CALMJS_WEBPACK_LOADERPLUGINS
 
 from calmjs.testing import utils
 from calmjs.testing.mocks import StringIO
 
-from calmjs.webpack.testing.utils import cls_setup_webpack_example_package
-from calmjs.webpack.testing.utils import generate_example_bundles
-from calmjs.webpack.testing.utils import skip_full_toolchain_test
+from calmjs.webpack.testing.utils import (
+    cls_setup_webpack_example_package,
+    cls_setup_webpack_loader_integration_packages,
+    generate_example_bundles,
+    skip_full_toolchain_test,
+)
 
 
 def run_node(src, *requires):
@@ -69,6 +74,28 @@ def _setup_extra_install(working_dir, packages):
         driver.pkg_manager_install(packages, production=False, merge=True)
 
 
+class RegistryRegisteredTests(unittest.TestCase):
+
+    def test_registry_registered(self):
+        # ensure that the registries are actually working
+        self.assertTrue(isinstance(
+            get_registry('calmjs.webpack.loaderplugins'),
+            loaderplugin.AutogenWebpackLoaderPluginRegistry,
+        ))
+        self.assertTrue(isinstance(
+            get_registry('calmjs.webpack.static.loaderplugins'),
+            LoaderPluginRegistry,
+        ))
+        self.assertTrue(isinstance(
+            get_registry('calmjs.module.webpackloader'),
+            loaderplugin.WebpackModuleLoaderRegistry,
+        ))
+        self.assertTrue(isinstance(
+            get_registry('calmjs.module.tests.webpackloader'),
+            loaderplugin.WebpackModuleLoaderRegistry,
+        ))
+
+
 @unittest.skipIf(*skip_full_toolchain_test())
 class ToolchainIntegrationTestCase(unittest.TestCase):
     """
@@ -98,10 +125,15 @@ class ToolchainIntegrationTestCase(unittest.TestCase):
         # integration harness will stub out the root distribution which
         # will break the installation of real tools.
         utils.setup_class_integration_environment(cls)
-        # also our test data.
+        # plus the integration setup for the loader packages
+        cls_setup_webpack_loader_integration_packages(cls)
+        # also the test data. note that after this stage, the module
+        # registry should NOT be reinitiated (at least until helpers for
+        # specific selection of which module entries are used is added).
         cls_setup_webpack_example_package(cls)
         # plus the extra packages
-        _setup_extra_install(cls._cls_tmpdir, ['example.loader'])
+        _setup_extra_install(cls._cls_tmpdir, [
+            'example.loader', 'example.styledemo'])
 
         # since our configuration paths will be at arbitrary locations
         # (i.e. temporary directories), NODE_PATH must be defined.
@@ -496,6 +528,65 @@ class ToolchainIntegrationTestCase(unittest.TestCase):
         self.assertIn(
             "no valid node_modules found - webpack may fail to locate itself",
             log)
+
+    def test_webpack_toolchain_webpackloader_style_css(self):
+        build_dir = utils.mkdtemp(self)
+        loaderplugin_sourcepath_maps = {
+            # just start off with one standalone file without imports
+            # for the most basic test to have one base test case;
+            # naturally ensure the name is available but mapped to the
+            # other @import free file.
+            'style': {
+                'style!css!example/styledemo/index.css': join(
+                    self._es_root, 'extra.css'),
+            },
+            'css': {
+                'css!example/styledemo/index.css': join(
+                    self._es_root, 'extra.css'),
+            },
+        }
+        export_target = join(utils.mkdtemp(self), 'example.styledemo.js')
+        webpack = toolchain.WebpackToolchain(
+            node_path=join(self._env_root, 'node_modules'))
+        spec = Spec(
+            transpile_sourcepath={
+                'example/styledemo/index': join(self._es_root, 'index.js'),
+            },
+            bundle_sourcepath={},
+            loaderplugin_sourcepath_maps=loaderplugin_sourcepath_maps,
+            export_target=export_target,
+            build_dir=build_dir,
+            calmjs_webpack_modname_loader_map={
+                'example/styledemo/index.css': ['style', 'css'],
+            },
+            webpack_output_library='example.styledemo',
+        )
+        with pretty_logging(stream=StringIO()) as stream:
+            webpack(spec)
+
+        self.assertIn(
+            "AutogenWebpackLoaderPluginRegistry registry "
+            "'calmjs.webpack.loaderplugins' generated loader handler 'style'",
+            stream.getvalue(),
+        )
+        self.assertIn(
+            "AutogenWebpackLoaderPluginRegistry registry "
+            "'calmjs.webpack.loaderplugins' generated loader handler 'css'",
+            stream.getvalue(),
+        )
+
+        self.assertTrue(exists(export_target))
+
+        with open(export_target) as fd:
+            # ensure the css is included
+            self.assertIn('.table { border: 1px solid #000; }', fd.read())
+
+        # This bundle will not actually run; for the actual testing, the
+        # calmjs.dev integration tests provide it.
+        stdout, stderr = run_node("""
+        var artifact = %s;
+        """, export_target)
+        self.assertNotEqual(stderr, '')
 
     # Tests using the Toolchain with the cli abstraction.
 
@@ -1317,7 +1408,7 @@ class ToolchainIntegrationTestCase(unittest.TestCase):
 
 
 @unittest.skipIf(karma is None, 'calmjs.dev or its karma module not available')
-class KarmatoolchainIntegrationTestCase(unittest.TestCase):
+class KarmaToolchainIntegrationTestCase(unittest.TestCase):
     """
     Test out the karma toolchain, involving webpack completely along
     with the karma testing framework as defined by calmjs.dev
@@ -1349,10 +1440,15 @@ class KarmatoolchainIntegrationTestCase(unittest.TestCase):
         # integration harness will stub out the root distribution which
         # will break the installation of real tools.
         utils.setup_class_integration_environment(cls)
-        # also our test data.
+        # plus the integration setup for the loader packages
+        cls_setup_webpack_loader_integration_packages(cls)
+        # also the test data. note that after this stage, the module
+        # registry should NOT be reinitiated (at least until helpers for
+        # specific selection of which module entries are used is added).
         cls_setup_webpack_example_package(cls)
         # plus the extra packages
-        _setup_extra_install(cls._cls_tmpdir, ['example.loader'])
+        _setup_extra_install(cls._cls_tmpdir, [
+            'example.loader', 'example.styledemo'])
 
     @classmethod
     def tearDownClass(cls):
@@ -1385,10 +1481,12 @@ class KarmatoolchainIntegrationTestCase(unittest.TestCase):
         utils.stub_stdouts(self)
         current_dir = utils.mkdtemp(self)
         build_dir = utils.mkdtemp(self)
+        report_dir = utils.mkdtemp(self)
         export_target = join(current_dir, 'example_package.js')
         with self.assertRaises(SystemExit) as e:
             runtime.main([
                 'karma', '--coverage',
+                '--cover-report-dir=' + report_dir,
                 'webpack', 'example.package',
                 '--export-target=' + export_target,
                 '--build-dir=' + build_dir,
@@ -1397,7 +1495,7 @@ class KarmatoolchainIntegrationTestCase(unittest.TestCase):
         self.assertTrue(exists(export_target))
 
         # verify that the coverage was recorded only for packge
-        coverage_file = join(self._env_root, 'coverage', 'coverage.json')
+        coverage_file = join(report_dir, 'coverage.json')
         with codecs.open(coverage_file, encoding='utf8') as fd:
             coverage = json.load(fd)
         expected = {
@@ -1411,10 +1509,12 @@ class KarmatoolchainIntegrationTestCase(unittest.TestCase):
         utils.stub_stdouts(self)
         current_dir = utils.mkdtemp(self)
         build_dir = utils.mkdtemp(self)
+        report_dir = utils.mkdtemp(self)
         export_target = join(current_dir, 'example_package.js')
         with self.assertRaises(SystemExit) as e:
             runtime.main([
                 'karma', '--coverage', '--cover-test',
+                '--cover-report-dir=' + report_dir,
                 'webpack', 'example.package',
                 '--export-target=' + export_target,
                 '--build-dir=' + build_dir,
@@ -1423,7 +1523,7 @@ class KarmatoolchainIntegrationTestCase(unittest.TestCase):
         self.assertTrue(exists(export_target))
 
         # verify that the coverage was recorded for test also.
-        coverage_file = join(self._env_root, 'coverage', 'coverage.json')
+        coverage_file = join(report_dir, 'coverage.json')
         with codecs.open(coverage_file, encoding='utf8') as fd:
             coverage = json.load(fd)
         expected = {
@@ -1450,10 +1550,12 @@ class KarmatoolchainIntegrationTestCase(unittest.TestCase):
         utils.stub_stdouts(self)
         current_dir = utils.mkdtemp(self)
         build_dir = utils.mkdtemp(self)
+        report_dir = utils.mkdtemp(self)
         export_target = join(current_dir, 'example_extras.js')
         with self.assertRaises(SystemExit) as e:
             runtime.main([
                 'karma', '--coverage', '--cover-test',
+                '--cover-report-dir=' + report_dir,
                 'webpack', 'example.extras',
                 '--export-target=' + export_target,
                 '--build-dir=' + build_dir,
@@ -1462,7 +1564,7 @@ class KarmatoolchainIntegrationTestCase(unittest.TestCase):
         self.assertTrue(exists(export_target))
 
         # verify that the coverage was recorded.
-        coverage_file = join(self._env_root, 'coverage', 'coverage.json')
+        coverage_file = join(report_dir, 'coverage.json')
         with codecs.open(coverage_file, encoding='utf8') as fd:
             coverage = json.load(fd)
         expected = {
@@ -1529,6 +1631,91 @@ class KarmatoolchainIntegrationTestCase(unittest.TestCase):
                 '-t', 'calmjs.webpack',
                 'example.extras',
             ])
+        self.assertEqual(e.exception.args[0], 0)
+
+    def test_module_rule_loaders_proper_assignment(self):
+        """
+        Since it is rather infeasible to test for styling changes with
+        just Node.js, here the actual karma/sinon tests encoded by the
+        test set up will be executed to show that colour changes are
+        applied by the styles encoded inside the artifact.
+        """
+
+        utils.stub_stdouts(self)
+        current_dir = utils.mkdtemp(self)
+        export_target = join(current_dir, 'example_styledemo.js')
+        with self.assertRaises(SystemExit) as e:
+            runtime.main([
+                'karma',
+                # note that the following --test-registry is required
+                # for calmjs-dev<2.2.0, as it does not know about the
+                # correct way to apply test loader repositories until
+                # then.
+                # '--test-registry=calmjs.module.simulated.tests,'
+                # 'calmjs.module.simulated.tests.webpackloader',
+                'webpack', 'example.styledemo',
+                '--export-target=' + export_target,
+            ])
+        self.assertEqual(e.exception.args[0], 0)
+        self.assertTrue(exists(export_target))
+
+        with open(export_target) as fd:
+            # ensure that the contents source by test registries are not
+            # encoded into the final artifact also
+            self.assertNotIn('hello world', fd.read())
+
+    def test_module_rule_loaders_proper_assignment_with_coverage(self):
+        """
+        Same as above, but ensure that coverage is generated
+        """
+
+        utils.stub_stdouts(self)
+        build_dir = utils.mkdtemp(self)
+        current_dir = utils.mkdtemp(self)
+        export_target = join(current_dir, 'example_styledemo.js')
+        with self.assertRaises(SystemExit) as e:
+            runtime.main([
+                'karma', '--coverage', '--cover-test',
+                '--cover-report-dir=' + build_dir,
+                'webpack', 'example.styledemo',
+                '--export-target=' + export_target,
+                '--build-dir=' + build_dir,
+            ])
+        self.assertEqual(e.exception.args[0], 0)
+        self.assertTrue(exists(export_target))
+
+        with open(join(build_dir, 'lcov', 'lcov.info')) as fd:
+            report = fd.read()
+            self.assertIn('index.js', report)
+            self.assertIn('styledemo', report)
+
+    def test_karma_test_runner_standalone_artifact_with_test_loaders(self):
+        """
+        Show that the standalone artifact won't conflict with the
+        test loaders also.
+        """
+
+        utils.stub_stdouts(self)
+        current_dir = utils.mkdtemp(self)
+        export_target = join(current_dir, 'example_styledemo.js')
+        with self.assertRaises(SystemExit) as e:
+            runtime.main([
+                'webpack', 'example.styledemo',
+                '--export-target=' + export_target,
+            ])
+
+        self.assertTrue(exists(export_target))
+
+        # leverage the karma run command to run the tests provided by
+        # the example.package against the resulting artifact.
+        with self.assertRaises(SystemExit) as e:
+            runtime.main([
+                'karma', 'run',
+                '--test-with-package', 'example.styledemo',
+                '--artifact', export_target,
+                '--toolchain-package', 'calmjs.webpack',
+            ])
+        # tests should pass against the resultant bundle
         self.assertEqual(e.exception.args[0], 0)
 
     def test_calmjs_artifact_test_verification(self):